% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{clinical}
\alias{clinical}
\alias{clinical.wide}
\title{Data set on ...
A simulated dataset, mimicking the study performed by Eichelsheim et al. (2011) who investigated whether there are
differences in patterns of negativity between families with and without an adolescent with externalizing problem
behavior. The problematic and nonproblematic group consist of 120 and 153 four-person families, respectively.
This dataset contains a measures of negativity for each of the 12 relationships.
Four roles are present: Mothers "M", fathers "F", the asolescent with externalizing problem behavior "T", and the adolescent sibling without problem behavior "S".
A wide version of the same data set is in \code{two.groups.wide}.}
\usage{
data(clinical)
}
\description{
The variables are as follows:
}
\details{
\itemize{
\item family.id An indicator for the family.
\item actor.id An indicator for the perceiver, either "M", "F", "T", or "S".
\item partner.id An indicator for the target, either "M", "F", "T", or "S".
\item neg Negativity measure.
}
}
\keyword{datasets}

