\name{equalMeans}
\alias{equalMeans}
\title{Test actor and partner means for equality}
\usage{
equalMeans(x, digits = 3)
}
\arguments{
  \item{x}{A fSRM object-}

  \item{digits}{Digits to which the printed results are
  rounded}
}
\description{
Test actor and partner means for equality, using a Wald
test.
}

