\name{import}
\alias{import}
\title{Import data sets from different formats}
\usage{
import()
}
\description{
Imports data sets from different formats and helps to
restructure them
}
\details{
TODO: Decribe some details of the functions
}
\examples{
# The result is saved in a variable called \\code{MyData}
import()
}

