\name{predict.fSRM}
\alias{predict.fSRM}
\title{Predict new cases based on a fitted fSRM model}
\usage{
  \method{predict}{fSRM} (object, newdata, ...)
}
\arguments{
  \item{object}{A fSRM object.}

  \item{newdata}{A data frame with exactly the same
  structure as the data frame on which the original fSRM
  object is based on.}

  \item{...}{Other parameters (currently not used)}
}
\description{
  Predict new cases based on a fitted fSRM model
}

