\docType{data}
\name{two.groups}
\alias{two.groups}
\title{Data set on negative interactions}
\format{A data frame with 2496 rows and 4 variables (208 families with 4 members each, round-robin design)}
\description{
A dataset containing ... TODO. Four roles are present:
Mothers "m", fathers "f", the older child "c", and the
younger child "y".
}
\details{
The variables are as follows:

\itemize{ \item family.id An indicator for the family.
\item actor.id An indicator for the perceiver, either "m",
"f", "c", or "y" \item partner.id An indicator for the
target, either "m", "f", "c", or "y" \item dep1 TODO:
Describe dependency measure.  \item dep2 TODO: Describe
dependency measure. }
}
\examples{
data(two.groups)
head(two.groups)
}
\references{
Cook, W. L. (2000). Understanding attachment security in
family context. \emph{Journal of Personality and Social
Psychology, 78}, 285–294. doi:10.1037/0022-3514.78.2.285
}
\keyword{datasets}

