\name{update.fSRM}
\alias{update.fSRM}
\title{Rerun a fSRM model with new parameters}
\usage{
\method{update}{fSRM}(object, evaluate = TRUE, ...)
}
\arguments{
  \item{object}{A fSRM object.}

  \item{evaluate}{Set to TRUE.}

  \item{...}{Other parameters (currently not used)}
}
\description{
Rerun a fSRM model with new parameters
}

